requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.5
tavily-python>=0.1.8
python-dotenv>=1.0.0
plotly>=5.18.0
boto3>=1.28.38
httpx>=0.24.1
fastapi>=0.104.0
uvicorn>=0.23.2 